Fixed number inaccuracy and negation misbehaviour

+++ src/compiler.zig
@@ -44,7 +44,7 @@ pub const Compiler = struct {
         self.getCurrentChunk().writeOpCode(op_code, line) catch {};
     }
 
-    pub fn emitValue(self: *Self, val: f16, line: u32) !void {
+    pub fn emitValue(self: *Self, val: f64, line: u32) !void {
         self.emitOpCodes(.op_value, try self.makeValue(val), line);
     }
 
@@ -53,7 +53,7 @@ pub const Compiler = struct {
         self.getCurrentChunk().writeByte(op2, line) catch {};
     }
 
-    pub fn makeValue(self: *Self, val: f16) !u8 {
+    pub fn makeValue(self: *Self, val: f64) !u8 {
         const value = try self.getCurrentChunk().addValue(val);
         if (value > 255) {
             std.debug.print("", .{}); //TODO
